+++ gora-cassandra-cql/src/main/java/org/apache/gora/cassandra/serializers/CassandraSerializer.java
@@ -26,7 +26,6 @@ import org.apache.gora.cassandra.store.CassandraClient;
 import org.apache.gora.cassandra.store.CassandraMapping;
 import org.apache.gora.cassandra.store.CassandraStore;
 import org.apache.gora.persistency.Persistent;
-import org.apache.gora.persistency.impl.PersistentBase;
 import org.apache.gora.query.Query;
 import org.apache.gora.query.Result;
 import org.apache.gora.store.DataStore;
@@ -34,7 +33,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -48,21 +46,14 @@ public abstract class CassandraSerializer<K, T extends Persistent> {
 
   protected Class<T> persistentClass;
   protected CassandraMapping mapping;
-  CassandraClient client;
-  private Map<String, String> userDefineTypeMaps;
-  private Schema persistentSchema;
+  protected CassandraClient client;
+  protected Map<String, String> userDefineTypeMaps;
 
-  CassandraSerializer(CassandraClient cc, Class<K> keyClass, Class<T> persistantClass, CassandraMapping mapping, Schema schema) {
+  CassandraSerializer(CassandraClient cc, Class<K> keyClass, Class<T> persistantClass, CassandraMapping mapping) {
     this.keyClass = keyClass;
     this.persistentClass = persistantClass;
     this.client = cc;
     this.mapping = mapping;
-    persistentSchema = schema;
-    try {
-      analyzePersistent();
-    } catch (Exception e) {
-      throw new RuntimeException("Error occurred while analyzing the persistent class, :" + e.getMessage());
-    }
   }
 
   /**
@@ -85,40 +76,17 @@ public abstract class CassandraSerializer<K, T extends Persistent> {
         break;
       case NATIVE:
       default:
-        serializer = new NativeSerializer(cc, dataStore.getKeyClass(), dataStore.getPersistentClass(), mapping, schema);
+        serializer = new NativeSerializer(cc, dataStore.getKeyClass(), dataStore.getPersistentClass(), mapping);
     }
     return serializer;
   }
 
-  private void analyzePersistent() throws Exception {
-    userDefineTypeMaps = new HashMap<>();
-    for (Field field : mapping.getFieldList()) {
-      String fieldType = field.getType();
-      if (fieldType.contains("frozen")) {
-        String udtType = fieldType.substring(fieldType.indexOf("<") + 1, fieldType.indexOf(">"));
-        if (this instanceof AvroSerializer) {
-          if (PersistentBase.class.isAssignableFrom(persistentClass)) {
-            Schema fieldSchema = persistentSchema.getField(field.getFieldName()).schema();
-            if (fieldSchema.getType().equals(Schema.Type.UNION)) {
-              for (Schema currentSchema : fieldSchema.getTypes()) {
-                if (currentSchema.getType().equals(Schema.Type.RECORD)) {
-                  fieldSchema = currentSchema;
-                  break;
-                }
-              }
-            }
-            String createQuery = CassandraQueryFactory.getCreateUDTTypeForAvro(mapping, udtType, fieldSchema);
-            userDefineTypeMaps.put(udtType, createQuery);
-          } else {
-            throw new RuntimeException("Unsupported Class for User Define Types, Please use PersistentBase class. field : " + udtType);
-          }
-        } else {
-          String createQuery = CassandraQueryFactory.getCreateUDTTypeForNative(mapping, persistentClass, udtType, field.getFieldName());
-          userDefineTypeMaps.put(udtType, createQuery);
-        }
-      }
-    }
-  }
+  /**
+   * In this method persistent class been analyzed to find inner records with UDT type, this method should call in every Cassandra serialization Constructor.
+   *
+   * @throws Exception
+   */
+  protected abstract void analyzePersistent() throws Exception;
 
 
   public void createSchema() {
@@ -166,27 +134,55 @@ public abstract class CassandraSerializer<K, T extends Persistent> {
     return fields.toArray(new String[0]);
   }
 
+  /**
+   * Inserts the persistent Object
+   *
+   * @param key   key value
+   * @param value persistent value
+   */
   public abstract void put(K key, T value);
 
+  /**
+   * Retrieves the persistent value according to the key
+   *
+   * @param key key value
+   * @return persistent value
+   */
   public abstract T get(K key);
 
+  /**
+   * Deletes persistent value according to the key
+   *
+   * @param key key value
+   * @return isDeleted
+   */
   public abstract boolean delete(K key);
 
+  /**
+   * Retrieves the persistent value according to the key and fields
+   *
+   * @param key    key value
+   * @param fields fields
+   * @return persistent value
+   */
   public abstract T get(K key, String[] fields);
 
+  /**
+   * Executes the given query and returns the results.
+   *
+   * @param dataStore Cassandra data store
+   * @param query     Cassandra Query
+   * @return Cassandra Result
+   */
   public abstract Result<K, T> execute(DataStore<K, T> dataStore, Query<K, T> query);
 
-  public boolean updateByQuery(Query query) {
-    List<Object> objectArrayList = new ArrayList<>();
-    String cqlQuery = CassandraQueryFactory.getUpdateByQuery(mapping, query, objectArrayList, persistentSchema);
-    ResultSet results;
-    if (objectArrayList.size() == 0) {
-      results = client.getSession().execute(cqlQuery);
-    } else {
-      results = client.getSession().execute(cqlQuery, objectArrayList.toArray());
-    }
-    return results.wasApplied();
-  }
+  /**
+   * Update the persistent objects
+   *
+   * @param query Cassandra Query
+   * @return isUpdated
+   */
+  public abstract boolean updateByQuery(Query query);
 
   public long deleteByQuery(Query query) {
     List<Object> objectArrayList = new ArrayList<>();
